+++ spring-security-advanced-authentication-ui/src/main/java/software/xdev/spring/security/web/authentication/ui/advanced/filters/AdvancedSharedPageGeneratingFilter.java
@@ -29,9 +29,9 @@ public interface AdvancedSharedPageGeneratingFilter<S extends AdvancedSharedPage
 {
 	List<String> DEFAULT_HEADER_ELEMENTS = List.of(
 		"<link "
-			+ "href=\"https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css\" "
+			+ "href=\"https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/css/bootstrap.min.css\" "
 			+ "rel=\"stylesheet\" "
-			+ "integrity=\"sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH\" "
+			+ "integrity=\"sha384-sRIl4kxILFvY47J16cr9ZwB07vP4J8+LH7qKQnuqkuIAvNWLzeN8tE5YBujZqJLB\" "
 			+ "crossorigin=\"anonymous\"/>");
 	
 	S setHeaderElements(List<String> headerElements);